chore: revert mongodb instrumentation move
